added: changed table structure, added id as long for easy search

+++ src/main/java/in/jewelx/jewelxbackend/dto/user/UserBrandResponseDto.java
@@ -10,6 +10,6 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserBrandResponseDto {
-	private int brandId;
+	private long brandId;
 	private String brandName;
 }
